+++ src/Atem.Core/Graphics/Timing/HDMA.cs
@@ -1,7 +1,7 @@
 ﻿using System.IO;
 using Atem.Core.State;
 
-namespace Atem.Core.Graphics
+namespace Atem.Core.Graphics.Timing
 {
     /// <summary>
     /// Implements Horizontal-Blank Direct Memory Access (HDMA) for the Game Boy Color.
